consolidate interface, move logic to start script
