+++ packages/express/src/index.ts
@@ -1,4 +1,4 @@
 export { default as Entangled } from '@entangled/interface';
 export { Service } from './interface';
 export { getContext } from './async_hook';
-export { default } from "./bootstrap";
+export { default } from "./start";
