+++ packages/express/src/interface.ts
@@ -1,8 +1,6 @@
 import Entangled from '@entangled/interface';
-import cookieParser from 'cookie-parser';
-import express, { Express, json, Router } from 'express';
+import { json, Router } from 'express';
 
-import { origin } from './gates';
 import { createNamespace, createRoute } from './routes';
 
 export class Service<R extends Entangled.Schema> {
@@ -28,27 +26,5 @@ export class Service<R extends Entangled.Schema> {
     const routes = Router().use(json());
     createRoute(routes, this.logic);
     return routes;
-  }
-
-  listen(
-    port: number, 
-    callback?: (app: Express) => void){
-
-    const app = express();
-  
-    app.use(origin());
-    app.use(this.routes());
-    app.use(cookieParser());
-
-    if(callback)
-      app.listen(port, () => callback(app));
-    else
-      app.listen(port);
-    
-    return app;
-  }
-
-  apply(to: Express | Router, root?: string){
-    createRoute(to, this.logic, root)
   }
 }
